実装しました

+++ resources/views/posts/index.blade.php
@@ -13,35 +13,37 @@
         {!! Form::close() !!}
     </div>
 
-    {{-- 投稿一覧表示 --}}
+
     <div>
         @foreach ($list as $list)
             {{-- as は、ループ内で現在の要素をどのように扱うかを指定 --}}
 
-            {{-- <img src="{{ asset('storage/' . $user->images) }}"> --}}
+            @if ($list->user_id == Auth::id() || Auth::user()->isFollowing($list->user_id))
+                {{-- フィルタリング条件を追加 --}}
+                <div class="icon-container">
+                    <!-- 投稿を編集するための記述  -->
+                    <a class="js-modal-open" href="" post="{{ $list->post }}" post_id="{{ $list->id }}">
+                        <img class="editicon" src="images/edit.png"></a>
 
-            <div class="icon-container">
-                <!-- 投稿を編集するための記述  -->
-                <a class="js-modal-open" href="" post="{{ $list->post }}" post_id="{{ $list->id }}">
-                    <img class="editicon" src="images/edit.png"></a>
-
-                {{-- 投稿を削除するための記述 --}}
-                <a class="trah-btn" href="/post/{{ $list->id }}/delete"
-                    onclick="return confirm('この投稿を削除します。よろしいでしょうか？')">
-                    <img class="trashicon" src="images/trash.png"></a>
-                {{-- aタグは基本get送信 --}}
+                    {{-- 投稿を削除するための記述 --}}
+                    <a class="trah-btn" href="/post/{{ $list->id }}/delete"
+                        onclick="return confirm('この投稿を削除します。よろしいでしょうか？')">
+                        <img class="trashicon" src="images/trash.png"></a>
+                    {{-- aタグは基本get送信 --}}
 
 
-            </div>
+                </div>
 
-            <ul>
-                {{-- postテーブルとuserテーブルから情報を取得 --}}
-                <li class="username">{{ $list->user->username }}</li>
-                <li>{{ $list->post }}</li>
-                <li>{{ $list->created_at }}</li>
-                <li> <img class="redicon" src="{{ asset('storage/' . $list->user->images) }}"></li>
-                {{-- ＄list-> post $list->user->username --}}
-            </ul>
+                {{-- 投稿一覧表示 --}}
+                <ul>
+                    {{-- postテーブルとuserテーブルから情報を取得 --}}
+                    <li class="username">{{ $list->user->username }}</li>
+                    <li>{{ $list->post }}</li>
+                    <li>{{ $list->created_at }}</li>
+                    <li> <img class="redicon" src="{{ asset('storage/' . $list->user->images) }}"></li>
+                    {{-- ＄list-> post $list->user->username --}}
+                </ul>
+            @endif {{-- フィルタリング条件の終了 --}}
         @endforeach
     </div>
 @endsection
